Altered load_students and save_students to accept user input, need to alter try_load_students to work with new load_students in directory.rb

+++ directory.rb
@@ -75,15 +75,23 @@ def main_menu(selection)
 end
 
 def save_students
+  #get input from user to choose a file to write to
+  puts "Please enter a file name including its type"
+  filename = STDIN.gets.chomp
   #open the file for writing
-  file = File.open("students.csv", "w")
-  #iterate over the array of students
-  @students.each do |student| 
-    student_data = [student[:name], student[:cohort]]
-    csv_line = student_data.join(",")
-    file.puts csv_line
+  if File.exist?(filename)
+    file = File.open("students.csv", "w")
+    #iterate over the array of students
+    @students.each do |student| 
+      student_data = [student[:name], student[:cohort]]
+      csv_line = student_data.join(",")
+      file.puts csv_line
+    end
+    file.close
+  else
+    puts "Sorry, #{filename} doesn't exist."
+    exit # quit the program
   end
-  file.close
 end
 
 def load_students()
